requirements.txt
pyfiglet==0.8.post1
python2==1.2
PyYAML==5.3.1
